+++ src/common/struct/DictionaryList.ts
