+++ src/static/images/markers/poi_fp.png
