+++ src/static/images/markers/poi_vld.png
