+++ src/static/images/markers/vehicles/plane.png
