+++ src/static/images/markers/vehicles/plane_red.png
