+++ src/static/images/markers/vehicles/rover.png
