+++ src/static/images/markers/vehicles/rover_red.png
