+++ src/static/index.ts
