+++ src/static/vehicle.json
